+++ src/commands/files.ts
@@ -1,5 +1,6 @@
 import { getArgList } from '../utils/args';
 import { spawnGitWithColor, parseStdoutByLine } from '../utils/commands';
+import { isStashRef } from '../utils/git';
 
 export async function files() {
   const args = getArgList();
@@ -12,7 +13,7 @@ export async function files() {
 }
 
 export async function getFiles(ref: string): Promise<string[]> {
-  if (/^stash@\{\d+\}$/.test(ref)) {
+  if (isStashRef(ref)) {
     return await parseStdoutByLine(`git stash show -p --include-untracked --name-only ${ref}`);
   } else {
     return await parseStdoutByLine(`git show --name-only --pretty=format:"" ${ref}`);